+++ packages/decap-cms-default-exports/src/index.js
@@ -13,7 +13,6 @@ import EmotionStyled from '@emotion/styled';
 import Immutable from 'immutable';
 import ImmutablePropTypes from 'react-immutable-proptypes';
 import Lodash from 'lodash/lodash';
-import Moment from 'moment';
 import PropTypes from 'prop-types';
 import React from 'react';
 import ReactDOM from 'react-dom';
@@ -35,7 +34,6 @@ export const DecapCmsDefaultExports = {
   Immutable,
   ImmutablePropTypes,
   Lodash,
-  Moment,
   PropTypes,
   React,
   ReactDOM,