+++ packages/decap-cms-locales/src/zh_Hant/index.js
@@ -174,6 +174,7 @@ const zh_Hant = {
       },
       datetime: {
         now: '現在',
+        clear: '清除',
       },
     },
   },