+++ website/src/cms/cms.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import CMS from 'netlify-cms-app';
+import CMS from 'decap-cms-app';
 import dayjs from 'dayjs';
 import Prism from 'prismjs';
-import { CacheProvider } from '@emotion/core';
+import { CacheProvider } from '@emotion/react';
 import createCache from '@emotion/cache';
 
 import BlogPostTemplate from '../components/blog-post-template';