+++ website/src/components/community-channels-list.js
@@ -46,7 +46,7 @@ function CommunityChannelsList({ channels }) {
     <StyledCommunityChannelsList>
       {channels.map(({ title, description, url }, idx) => (
         <li key={idx}>
-          <a href={url} target="_blank" rel="noopener noreferrer">
+          <a href={url} rel="noopener noreferrer">
             <strong>{title}</strong>
             <p>{description}</p>
           </a>