+++ website/src/components/release.js
@@ -1,7 +1,7 @@
-import React from 'react';
+/** @jsx jsx */
+import { jsx, css } from '@emotion/react';
 import moment from 'moment';
 import styled from '@emotion/styled';
-import { css } from '@emotion/core';
 
 import Markdownify from '../components/markdownify';
 import theme from '../theme';