+++ website/src/pages/services.js
@@ -1,15 +1,18 @@
-import React from 'react';
+/** @jsx jsx */
+import { jsx, css } from '@emotion/react';
 import { Helmet } from 'react-helmet';
 import { graphql } from 'gatsby';
 import { trimStart, trimEnd } from 'lodash';
 
 import TwitterMeta from '../components/twitter-meta';
 import Layout from '../components/layout';
-import BlogPostTemplate from '../components/blog-post-template';
+import Container from '../components/container';
+import Markdown from '../components/markdown';
+import Page from '../components/page';
 
 function BlogPost({ data }) {
   const { html, frontmatter } = data.markdownRemark;
-  const { author, title, date, description, meta_description, image, canonical_url } = frontmatter;
+  const { title, description, meta_description, image } = frontmatter;
   const { siteUrl } = data.site.siteMetadata;
   const imageUrl = image && `${trimEnd(siteUrl, '/')}/${trimStart(image, '/')}`;
   const desc = description || meta_description;
@@ -20,11 +23,20 @@ function BlogPost({ data }) {
         <title>{title}</title>
         {desc && <meta name="description" content={desc} />}
         {image && <meta name="image" property="og:image" content={imageUrl} />}
-        {author && <meta name="author" content={author} />}
-        {canonical_url && <link rel="canonical" href={canonical_url} />}
       </Helmet>
       <TwitterMeta title={title} description={desc} image={imageUrl} />
-      <BlogPostTemplate title={title} author={author} date={date} html={html} />
+      <Container size="sm">
+        <Page as="article">
+          <h1
+            css={css`
+              margin-bottom: 0;
+            `}
+          >
+            {title}
+          </h1>
+          <Markdown html={html} />
+        </Page>
+      </Container>
     </Layout>
   );
 }
@@ -32,21 +44,18 @@ function BlogPost({ data }) {
 export default BlogPost;
 
 export const pageQuery = graphql`
-  query blogPost($slug: String!) {
+  query servicesPage {
     site {
       siteMetadata {
         siteUrl
       }
     }
-    markdownRemark(fields: { slug: { eq: $slug } }) {
+    markdownRemark(fileAbsolutePath: { regex: "/services/" }) {
       frontmatter {
         title
         description
         # meta_description
-        date(formatString: "MMMM D, YYYY")
-        author
         image
-        canonical_url
       }
       html
     }