+++ packages/decap-cms-lib-widgets/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.1-beta.0](https://github.com/decaporg/decap-cms/compare/decap-cms-lib-widgets@3.0.0...decap-cms-lib-widgets@3.0.1-beta.0) (2023-11-23)
+
+
+### Performance Improvements
+
+* replace moment with dayjs ([#6980](https://github.com/decaporg/decap-cms/issues/6980)) ([22370b1](https://github.com/decaporg/decap-cms/commit/22370b13e49a4a5f58a60ebd4bc40ce4b141eb11))
+
+
+
+
+
 # [3.0.0](https://github.com/decaporg/decap-cms/compare/decap-cms-lib-widgets@1.9.0...decap-cms-lib-widgets@3.0.0) (2023-08-18)
 
 **Note:** Version bump only for package decap-cms-lib-widgets