+++ packages/decap-cms/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.1.0-beta.3](https://github.com/decaporg/decap-cms/compare/decap-cms@3.1.0-beta.2...decap-cms@3.1.0-beta.3) (2023-11-23)
+
+**Note:** Version bump only for package decap-cms
+
+
+
+
+
+# [3.1.0-beta.2](https://github.com/decaporg/decap-cms/compare/decap-cms@3.1.0-beta.1...decap-cms@3.1.0-beta.2) (2023-11-23)
+
+**Note:** Version bump only for package decap-cms
+
+
+
+
+
+# [3.1.0-beta.1](https://github.com/decaporg/decap-cms/compare/decap-cms@3.1.0-beta.0...decap-cms@3.1.0-beta.1) (2023-11-15)
+
+**Note:** Version bump only for package decap-cms
+
+
+
+
+
 # [3.1.0-beta.0](https://github.com/decaporg/decap-cms/compare/decap-cms@3.1.0...decap-cms@3.1.0-beta.0) (2023-10-20)